Add files via upload
This file is used to create the dihe.txt file used by NAMD colvars for restraining lipid tails. This was created by charmm-gui but I altered it to make two different files that have the information for both sets of membranes. These 2 files need to be combined in to one file manually for use by colvars.